requirements.txt
python-dotenv==1.1.0
python-telegram-bot==22.0
Requests==2.32.3
selenium==4.31.0
